Oleg: finally fixed horse morphing bug... I'm a moron

+++ src/mesh.cpp
@@ -43,18 +43,18 @@ void gl::Mesh::Draw(const std::vector<glm::mat4>& modelMatrices, size_t transfor
     glBindVertexArray(vaoAndVbo[0]);
 
     // Update pointers here in case multiple models use the same VAO/VBO.
-    glEnableVertexAttribArray(transformModelOffset);
-    glVertexAttribPointer(transformModelOffset, 4, GL_FLOAT, GL_FALSE, 4 * 4 * sizeof(float), (void*)0);
-    glEnableVertexAttribArray(transformModelOffset + 1);
-    glVertexAttribPointer(transformModelOffset + 1, 4, GL_FLOAT, GL_FALSE, 4 * 4 * sizeof(float), (void*)(4 * sizeof(float)));
-    glEnableVertexAttribArray(transformModelOffset + 2);
-    glVertexAttribPointer(transformModelOffset + 2, 4, GL_FLOAT, GL_FALSE, 4 * 4 * sizeof(float), (void*)(2 * 4 * sizeof(float)));
-    glEnableVertexAttribArray(transformModelOffset + 3);
-    glVertexAttribPointer(transformModelOffset + 3, 4, GL_FLOAT, GL_FALSE, 4 * 4 * sizeof(float), (void*)(3 * 4 * sizeof(float)));
-    glVertexAttribDivisor(transformModelOffset, 1);
-    glVertexAttribDivisor(transformModelOffset + 1, 1);
-    glVertexAttribDivisor(transformModelOffset + 2, 1);
-    glVertexAttribDivisor(transformModelOffset + 3, 1);
+    glEnableVertexAttribArray((unsigned int)transformModelOffset);
+    glVertexAttribPointer((unsigned int)transformModelOffset, 4, GL_FLOAT, GL_FALSE, 4 * 4 * sizeof(float), (void*)0);
+    glEnableVertexAttribArray((unsigned int)transformModelOffset + 1);
+    glVertexAttribPointer((unsigned int)transformModelOffset + 1, 4, GL_FLOAT, GL_FALSE, 4 * 4 * sizeof(float), (void*)(4 * sizeof(float)));
+    glEnableVertexAttribArray((unsigned int)transformModelOffset + 2);
+    glVertexAttribPointer((unsigned int)transformModelOffset + 2, 4, GL_FLOAT, GL_FALSE, 4 * 4 * sizeof(float), (void*)(2 * 4 * sizeof(float)));
+    glEnableVertexAttribArray((unsigned int)transformModelOffset + 3);
+    glVertexAttribPointer((unsigned int)transformModelOffset + 3, 4, GL_FLOAT, GL_FALSE, 4 * 4 * sizeof(float), (void*)(3 * 4 * sizeof(float)));
+    glVertexAttribDivisor((unsigned int)transformModelOffset, 1);
+    glVertexAttribDivisor((unsigned int)transformModelOffset + 1, 1);
+    glVertexAttribDivisor((unsigned int)transformModelOffset + 2, 1);
+    glVertexAttribDivisor((unsigned int)transformModelOffset + 3, 1);
 
     glBindBuffer(GL_ARRAY_BUFFER, 0);
     glBindVertexArray(0);
